separated into own components
TodoList.js and Todo.js are separate from App.js inside the components folder
