+++ src/components/Todo.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-function Todo(props, i) {
+function Todo(props) {
   return <li>{props.todo}</li>;
 }
